python/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "sensitivity-surrogacy"
version = "0.0.1"
description = "Sensitivity analysis and partial identification using surrogate outcomes"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "scikit-learn",
    "python-glmnet",
    "econml",
    "quantile-forest",
    "xgboost",
]

[tool.setuptools.packages.find]
where = ["."]
